feat: A new feature
Replace "framename" with "_self | _parent | framename" inside "home.html" file to open linked document of "View" option in initial webpage.

This is done to see if link will open in same frame it is a default value and in parent frameset with named frame and see what result will be produced.